+++ CHANGELOG.rst
@@ -8,8 +8,11 @@ Version 0.10 (unreleased)
 
 * Addition of ``nearest`` and ``nearest_all`` functions to ``STRtree`` for
   GEOS >= 3.6 to find the nearest neighbors (#272).
+* Enable bulk construction of geometries with different number of coordinates
+  by optionally taking index arrays in the constructors ``multipoints``,
+  ``points``, ``linestrings``, and ``linearrings`` (#322).
 * Enable bulk construction of collections with different number of geometries
-  by optionally taking an index arrays in the constructors ``multipoints``,
+  by optionally taking index arrays in the constructors ``multipoints``,
   ``multilinestrings``, ``multipolygons``, and ``geometrycollections`` (#290).
 * Released GIL for ``points``, ``linestrings``, ``linearrings``, and 
   ``polygons`` (without holes) (#310).